Add cra-monitor daemon
Create a new simple daemon only dedicated to monitor CRAs.
Can be installed on any CRA, it will listen in default netNS and will
monitor the netNS in which it as been executed (via ip netns exec).
It will expose a http /metrics url that can be requested to monitor
netlink and frr data. Does not require any certificate to be requested.

Signed-off-by: Romain Delhomel <romain.delhomel@6wind.com>